+++ src/Wheelzy.Aplication/Interfaces/Repositories/QuoteProjection.cs
@@ -1,5 +1,5 @@
 ﻿namespace Wheelzy.Application.Interfaces.Repositories
 {
-    public readonly record struct QuoteProjection(long OrderBuyerQuoteId, decimal Amount);
+    public readonly record struct QuoteProjection(int OrderBuyerQuoteId, decimal Amount);
 
 }
