+++ src/Wheelzy.Aplication/Models/OrderBuyerQuote.cs
@@ -2,8 +2,8 @@
 {
     public class OrderBuyerQuote
     {
-        public long OrderBuyerQuoteId { get; set; }
-        public long OrderId { get; set; }
+        public int OrderBuyerQuoteId { get; set; }
+        public int OrderId { get; set; }
         public int BuyerZipCoverageId { get; set; }
         public decimal Amount { get; set; }
         public DateTime CreatedAt { get; set; }
